+++ Functions/processAvtopark.js
